requirements.txt
# ── Core ML / Compute ────────────────────────────────────────────────────────
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# TensorRT Python bindings (wheel must match TensorRT 8.6 installed in container)
tensorrt==8.6.1

# NVIDIA Management Library bindings
pynvml==11.5.0

# ── Computer Vision ───────────────────────────────────────────────────────────
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80

# YOLOv8 / Ultralytics
ultralytics==8.1.18

# ── ONNX ─────────────────────────────────────────────────────────────────────
onnx==1.15.0
onnxsim==0.4.35
onnxruntime-gpu==1.17.1

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.30.0
plotly==5.18.0

# ── Data / IO ─────────────────────────────────────────────────────────────────
pandas==2.2.0
pyarrow==15.0.0
numpy==1.26.4

# ── RL Environment ────────────────────────────────────────────────────────────
gymnasium==0.29.1

# ── Scientific ────────────────────────────────────────────────────────────────
scipy==1.12.0

# ── Config / Utilities ────────────────────────────────────────────────────────
pyyaml==6.0.1

# ── Testing / Dev ─────────────────────────────────────────────────────────────
pytest==8.0.2
pytest-cov==4.1.0
pytest-timeout==2.2.0
hypothesis==6.98.2
black==24.2.0
isort==5.13.2
mypy==1.8.0
